Show expanded speaker notes on print page
The `/print.html` page is concatenates all pages in the book and can
be reached from the printer icon in the top-right corner.

On this page, we will have many speaker notes. Since the page is for
people who want all information in one place, I think it makes sense
to render all these notes in their expanded form.

I wrote the logic for this as a separate little function since it has
to be different: we don’t want to add an `id` attribute to the
generated headers (since there will be more than one of them) and we
don’t want to generate links to open the notes into a separate window.