de: normalize with mdbook-i18n-helpers 0.2.2 (#1114)
Before, po/de.po had these statistics:

410 translated messages, 697 fuzzy translations, 674 untranslated
messages.

Afterwards, the statistics for po/de.po is:

556 translated messages, 814 fuzzy translations, 1097 untranslated
messages.

The number of translated messages stayed at 23%.

With this change, it becomes important to use the latest version of
mdbook-i18n-helpers when viewing the translation locally. To update to
the latest version, run

    cargo install mdbook-i18n-helpers

You will now be able to serve the translation locally.

Part of #330.