ru: normalize with mdbook-i18n-helpers 0.2.2 (#1110)
Before, po/ru.po had these statistics:

217 translated messages, 35 fuzzy translations, 1529 untranslated
messages.

Afterwards, the statistics for po/ru.po is:

374 translated messages, 47 fuzzy translations, 2046 untranslated
messages.

The number of translated messages changed from 12% to 15%.

With this change, it becomes important to use the latest version of
mdbook-i18n-helpers when viewing the translation locally. To update to
the latest version, run

    cargo install mdbook-i18n-helpers

You will now be able to serve the translation locally.

Part of #330.