es: normalize with mdbook-i18n-helpers 0.2.2 (#1109)
Before, po/es.po had these statistics:

386 translated messages, 133 fuzzy translations, 1262 untranslated
messages.

Afterwards, the statistics for po/es.po is:

585 translated messages, 142 fuzzy translations, 1740 untranslated
messages.

The number of translated messages changed from 22% to 24%.

With this change, it becomes important to use the latest version of
mdbook-i18n-helpers when viewing the translation locally. To update to
the latest version, run

    cargo install mdbook-i18n-helpers

You will now be able to serve the translation locally.

Part of #330.