tr: normalize with mdbook-i18n-helpers 0.2.2 (#1098)
Before, po/tr.po had these statistics:

164 translated messages, 6 fuzzy translations, 1611 untranslated
messages.

Afterwards, the statistics for po/tr.po is:

165 translated messages, 6 fuzzy translations, 2296 untranslated
messages.

The number of translated messages changed from 9% to 7%.

With this change, it becomes important to use the latest version of
mdbook-i18n-helpers when viewing the translation locally. To update to
the latest version, run

    cargo install mdbook-i18n-helpers

You will now be able to serve the translation locally.

Part of #330.