zh-CN: normalize with mdbook-i18n-helpers 0.2.2 (#1097)
Before, po/zh-CN.po had these statistics:

693 translated messages, 90 fuzzy translations, 998 untranslated
messages.

Afterwards, the statistics for po/zh-CN.po is:

1079 translated messages, 121 fuzzy translations, 1267 untranslated
messages.

The number of translated messages changed from 39% to 44%.

With this change, it becomes important to use the latest version of
mdbook-i18n-helpers when viewing the translation locally. To update to
the latest version, run

    cargo install mdbook-i18n-helpers

You will now be able to serve the translation locally.

Part of #330.