pt-BR: remove translated variable names (#1472)
With #1460 merged, we took away the possibility of translating variable,
struct, and function names in the code snippets.

This broke the Brazilian Portuguese translation because it _does_
translate those things quite frequently! This PR reverts the
translations back to the English original.

I tried to also update the surrounding messages to use the English
words: so when a slide talked about a `nome` variable before, it now
uses `name` instead. We don’t have automated tests for this, so I hope I
did this correctly.